requirements.txt
psutil==5.9.5
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
loguru==0.7.2
snmp-pysnmp==4.4.12
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1 